backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-multipart==0.0.6
python-dotenv==1.0.0
numpy==1.24.3
pandas==1.5.3
ibm-watson-machine-learning==1.0.327
google-generativeai==0.3.2
jinja2==3.1.2
pydantic==2.5.0
reportlab==4.0.7
Pillow==10.1.0
requests==2.31.0
aiofiles==23.2.1
pydantic-settings==2.1.0